requirements.txt
mysql-connector-python==9.0.0
python-dotenv==1.0.1
openai==1.51.0
pydantic==2.8.2
rich==13.7.1
